Fix implicit integer precision loss

+++ JSAnimatedImagesView.m
@@ -260,7 +260,7 @@ static const CGFloat JSAnimatedImagesViewImageViewsBorderOffset = 10;
         return [self randomNumberBetweenNumber:maxNumber andNumber:minNumber];
     }
     
-    NSUInteger randomInt = (arc4random_uniform(maxNumber - minNumber + 1)) + minNumber;
+    NSUInteger randomInt = (arc4random_uniform((u_int32_t)(maxNumber - minNumber + 1))) + minNumber;
     
     return randomInt;
 }
